requirements.txt
fastapi
pydantic
uvicorn
google-genai
requests
trafilatura
python-dotenv
